rockchip: rk3568: Update board.c file for RGxx3 to auto set panel
Update the board.c file for the Anbernic RGxx3 to autodetect the
panel for all 353 series devices, and update the devicetree
automatically if a newer revision of the panel is found.

Also redo the board struct to make it easier to add newer devices
in the future.

Note with this change the startup light is now red, turns orange
when the device autodetects the hardware and hands off to Linux,
and it is implied that Linux will change the color to green.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>